requirements.txt
amqp==2.5.2
asgiref==3.2.7
asn1crypto==0.24.0
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
click==7.1.2
coverage==5.1
cryptography==2.3
Cython==0.29.18
Django==3.0.6
django-celery-results==1.2.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.6
importlib-metadata==1.6.0
Jinja2==2.11.2
joblib==0.15.1
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.8
livereload==2.6.1
lunr==0.5.8
Markdown==3.2.2
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==5.2.0
mkdocs-material-extensions==1.0
nltk==3.5
psycopg2==2.8.5
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
pymdown-extensions==7.1
python-apt==1.6.5+ubuntu0.2
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
regex==2020.5.14
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.3.1
tornado==6.0.4
tqdm==4.46.0
unattended-upgrades==0.1
vine==1.3.0
zipp==3.1.0
